requirements.txt
groq==0.4.1
google-generativeai==0.3.2
openai==1.54.0
anthropic==0.39.0
